Fix middleware.Recoverer under Go 1.17+
Under Go 1.17+ the lines of the debug stack aren't guaranteed to
contain a period, notability a line can just read "panic".

This commit adds some defensive programming to ensure we don't
attempt to slice with a negative index and adds a very basic test
to catch the negative index panic without this change.