Add a check to prevent modifying outer scope routers
This check adds a panic that triggers when the build function passed to
(*Mux).Route() or (*Mux).Group() attempts to set routes or middleware on a
router is a parent of its argument router.

Example:

```
root := chi.NewRouter()
root.Route("/api", func(api chi.Router) {
    root.Use(DatabaseTransaction)
    // ...
})
root.Get("/health-check", HealthCheck)
```

This reads as if the DatabaseTransaction was intended to be used as middleware
on the api sub-router but it is actually used on the top-level router, making
it run in the hot path of the /health-check handler.

Adding this check will catch such mistakes.