Add ability to poison contexts, preventing reuse
This adds the ability to mark routing contexts as "poisoned", which
prevents them from being returned to the sync.Pool after request
handling.

This is useful in scenarios where the context may be held _after_ a
request has been served to a user, like asynchronous background tasks,
mirror testing, etc.